# By Jonas Pedersen
# Via Jonas Pedersen
* commit '4266ced3415835b8403cb145f73ea7afadd3f87f':
  [MTICKET-6348] Adds javadoc to public ticket methods.
  [MTICKET-6348] Cleans up unit tests.
  [MTICKET-6348] Adds a activate, cancel and refund function to ticket.
  [MTICKET-6348] Finalizes transitions and unit tests.
  [MTICKET-6348] Error state changes, now requires an error code.
  [MTICKET-6348] Adds a simple state machine.
  [MTICKET-6348] Start implementing ticket state machine.

+++ src/main/resources/db/migration/common/V201510121321__initial_configuration.sql
@@ -422,13 +422,13 @@ CREATE TABLE `ticket_property` (
 /*Table structure for table `ticket_state` */
 
 CREATE TABLE `ticket_state` (
-  `id`                INT(11)   NOT NULL AUTO_INCREMENT,
-  `ticket_state`      INT(11)   NOT NULL DEFAULT '0',
-  `transaction_state` INT(11)   NOT NULL DEFAULT '0',
-  `action`            INT(11)   NOT NULL DEFAULT '0',
-  `error_code`        INT(11)   NOT NULL DEFAULT '0',
-  `timestamp`         TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
-  `ticket_id`         INT(11)   NOT NULL,
+  `id`                BIGINT(20) NOT NULL AUTO_INCREMENT,
+  `ticket_state`      INT(11)    NOT NULL DEFAULT '0',
+  `transaction_state` INT(11)    NOT NULL DEFAULT '0',
+  `action`            INT(11)    NOT NULL DEFAULT '0',
+  `error_code`        INT(11)    NOT NULL DEFAULT '0',
+  `timestamp`         TIMESTAMP  NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
+  `ticket_id`         INT(11)    NOT NULL,
   PRIMARY KEY (`id`),
   KEY `ticket_state` (`id`, `ticket_id`),
   KEY `ticket_id_state` (`ticket_id`, `id`)
